@@ -1,44 +1,41 @@
 {
     "name": "guillaumebriday/laravel-blog",
-    "description": "Laravel 7.0 blog.",
+    "description": "Laravel 8.0 blog.",
     "keywords": ["framework", "laravel", "blog", "example"],
     "license": "MIT",
     "type": "project",
     "require": {
         "php": "^7.4",
-        "barryvdh/laravel-debugbar": "3.3.3",
-        "doctrine/dbal": "2.10.2",
-        "fideloper/proxy": "4.3.0",
-        "fruitcake/laravel-cors": "2.0.1",
-        "laravel/framework": "7.12.0",
-        "laravel/horizon": "4.3.3",
-        "laravel/socialite": "4.3.2",
-        "laravel/tinker": "2.4.1",
-        "laravel/ui": "2.0.3",
-        "laravelcollective/html": "6.1.2",
-        "nunomaduro/collision": "4.2.0",
-        "predis/predis": "1.1.1",
+        "barryvdh/laravel-debugbar": "3.5.1",
+        "doctrine/dbal": "2.11.1",
+        "fideloper/proxy": "4.4.0",
+        "fruitcake/laravel-cors": "2.0.2",
+        "laravel/framework": "8.9.0",
+        "laravel/horizon": "5.2.1",
+        "laravel/socialite": "5.0.1",
+        "laravel/tinker": "2.4.2",
+        "laravel/ui": "3.0.0",
+        "laravelcollective/html": "6.2.0",
+        "nunomaduro/collision": "5.0.2",
+        "predis/predis": "1.1.6",
         "pusher/pusher-php-server": "4.1.4",
-        "spatie/laravel-backup": "6.10.0",
+        "spatie/laravel-backup": "6.11.4",
         "spatie/laravel-medialibrary": "8.10.1"
     },
     "require-dev": {
-        "beyondcode/laravel-dump-server": "1.4.0",
         "filp/whoops": "2.7.3",
         "friendsofphp/php-cs-fixer": "2.16.4",
         "fzaninotto/faker": "1.9.1",
-        "laravel/homestead": "11.2",
-        "laravel/telescope": "3.5.1",
-        "mockery/mockery": "1.4.0",
-        "phpunit/phpunit": "9.1.5"
+        "laravel/homestead": "11.2.2",
+        "laravel/telescope": "4.1.0",
+        "mockery/mockery": "1.4.2",
+        "phpunit/phpunit": "9.4.0"
     },
     "autoload": {
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ],
         "psr-4": {
-            "App\\": "app/"
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
         },
         "files": [
             "app/Helpers/date.php"
