@@ -14,7 +14,7 @@ class CommentTest extends TestCase
 
     public function testCommentIndex()
     {
-        factory(Comment::class, 2)->create();
+        Comment::factory()->count(2)->create();
 
         $this->json('GET', '/api/v1/comments')
             ->assertOk()
@@ -46,9 +46,9 @@ class CommentTest extends TestCase
 
     public function testUsersComments()
     {
-        $user = factory(User::class)->create();
-        factory(Comment::class, 10)->create(['author_id' => $user->id]);
-        factory(Comment::class, 10)->create();
+        $user = User::factory()->create();
+        Comment::factory()->count(10)->create(['author_id' => $user->id]);
+        Comment::factory()->count(10)->create();
 
         $this->json('GET', "/api/v1/users/{$user->id}/comments")
             ->assertOk()
@@ -88,9 +88,9 @@ class CommentTest extends TestCase
 
     public function testPostsComments()
     {
-        $post = factory(Post::class)->create();
-        factory(Comment::class, 10)->create(['post_id' => $post->id]);
-        factory(Comment::class, 10)->create();
+        $post = Post::factory()->create();
+        Comment::factory()->count(10)->create(['post_id' => $post->id]);
+        Comment::factory()->count(10)->create();
 
         $this->json('GET', "/api/v1/posts/{$post->id}/comments")
             ->assertOk()
@@ -130,7 +130,7 @@ class CommentTest extends TestCase
 
     public function testStore()
     {
-        $post = factory(Post::class)->create();
+        $post = Post::factory()->create();
 
         $this->actingAsUser('api')
             ->json('POST', "/api/v1/posts/{$post->id}/comments", $this->validParams())
@@ -149,7 +149,7 @@ class CommentTest extends TestCase
 
     public function testCommentShow()
     {
-        $comment = factory(Comment::class)->create([
+        $comment = Comment::factory()->create([
             'content' => 'The Empire Strikes Back'
         ]);
 
@@ -192,7 +192,7 @@ class CommentTest extends TestCase
 
     public function testCommentDelete()
     {
-        $comment = factory(Comment::class)->create();
+        $comment = Comment::factory()->create();
 
         $this->actingAs($comment->author, 'api')
             ->json('DELETE', "/api/v1/comments/{$comment->id}")
@@ -211,7 +211,7 @@ class CommentTest extends TestCase
 
     public function testCommentDeleteUnauthorized()
     {
-        $comment = factory(Comment::class)->create();
+        $comment = Comment::factory()->create();
 
         $this->actingAsUser('api')
             ->json('DELETE', "/api/v1/comments/{$comment->id}")
@@ -223,7 +223,7 @@ class CommentTest extends TestCase
 
     public function testCommentsDeleteUnauthenticated()
     {
-        $comment = factory(Comment::class)->create();
+        $comment = Comment::factory()->create();
         $this->json('DELETE', "/api/v1/comments/{$comment->id}")
             ->assertUnauthorized()
             ->assertJson([
