@@ -15,7 +15,7 @@ class PostTest extends TestCase
 
     public function testPostIndex()
     {
-        factory(Post::class, 2)->create();
+        Post::factory()->count(2)->create();
 
         $this->json('GET', '/api/v1/posts')
             ->assertOk()
@@ -49,8 +49,8 @@ class PostTest extends TestCase
 
     public function testUsersPosts()
     {
-        $user = factory(User::class)->create();
-        factory(Post::class)->create(['author_id' => $user->id]);
+        $user = User::factory()->create();
+        Post::factory()->create(['author_id' => $user->id]);
 
         $this->json('GET', "/api/v1/users/{$user->id}/posts")
             ->assertOk()
@@ -84,8 +84,8 @@ class PostTest extends TestCase
 
     public function testUsersPostsFail()
     {
-        $user = factory(User::class)->create();
-        factory(Post::class)->create(['author_id' => $user->id]);
+        $user = User::factory()->create();
+        Post::factory()->create(['author_id' => $user->id]);
 
         $this->json('GET', '/api/v1/users/314/posts')
             ->assertNotFound()
@@ -96,11 +96,11 @@ class PostTest extends TestCase
 
     public function testPostShow()
     {
-        $post = factory(Post::class)->create([
+        $post = Post::factory()->create([
             'title' => 'The Empire Strikes Back',
             'content' => 'A Star Wars Story'
         ]);
-        factory(Comment::class, 2)->create(['post_id' => $post->id]);
+        Comment::factory()->count(2)->create(['post_id' => $post->id]);
 
         $this->json('GET', "/api/v1/posts/{$post->id}")
             ->assertOk()
@@ -139,7 +139,7 @@ class PostTest extends TestCase
 
     public function testUpdate()
     {
-        $post = factory(Post::class)->create();
+        $post = Post::factory()->create();
         $params = $this->validParams();
 
         $this->actingAsAdmin('api')
@@ -155,7 +155,7 @@ class PostTest extends TestCase
 
     public function testUpdateFail()
     {
-        $post = factory(Post::class)->create();
+        $post = Post::factory()->create();
 
         $this->actingAsUser('api')
             ->json('PATCH', "/api/v1/posts/{$post->id}", $this->validParams())
@@ -189,7 +189,7 @@ class PostTest extends TestCase
 
     public function testPostDelete()
     {
-        $post = factory(Post::class)->create();
+        $post = Post::factory()->create();
 
         $this->actingAsAdmin('api')
             ->json('DELETE', "/api/v1/posts/{$post->id}")
@@ -200,7 +200,7 @@ class PostTest extends TestCase
 
     public function testPostDeleteUnauthorized()
     {
-        $post = factory(Post::class)->create();
+        $post = Post::factory()->create();
 
         $this->actingAsUser('api')
             ->json('DELETE', "/api/v1/posts/{$post->id}")
