@@ -15,11 +15,11 @@ class PostTest extends TestCase
 
     public function testIndex()
     {
-        $anakin = factory(User::class)->states('anakin')->create();
+        $anakin = User::factory()->anakin()->create();
 
-        $post = factory(Post::class)->create(['author_id' => $anakin->id]);
-        factory(Post::class, 2)->create();
-        factory(Comment::class, 3)->create(['post_id' => $post->id]);
+        $post = Post::factory()->create(['author_id' => $anakin->id]);
+        Post::factory()->count(2)->create();
+        Comment::factory()->count(3)->create(['post_id' => $post->id]);
 
         $this->get('/')
             ->assertOk()
@@ -32,8 +32,8 @@ class PostTest extends TestCase
 
     public function testSearch()
     {
-        factory(Post::class, 3)->create();
-        $post = factory(Post::class)->create(['title' => 'Hello Obiwan']);
+        Post::factory()->count(3)->create();
+        $post = Post::factory()->create(['title' => 'Hello Obiwan']);
 
         $this->get('/?q=Hello')
             ->assertOk()
@@ -44,9 +44,9 @@ class PostTest extends TestCase
 
     public function testShow()
     {
-        $post = factory(Post::class)->create();
-        factory(Comment::class, 2)->create(['post_id' => $post->id]);
-        factory(Comment::class)->create(['post_id' => $post->id]);
+        $post = Post::factory()->create();
+        Comment::factory()->count(2)->create(['post_id' => $post->id]);
+        Comment::factory()->create(['post_id' => $post->id]);
 
         $this->actingAsUser()
             ->get("/posts/{$post->slug}")
@@ -60,7 +60,7 @@ class PostTest extends TestCase
 
     public function testShowUnauthenticated()
     {
-        $post = factory(Post::class)->create();
+        $post = Post::factory()->create();
 
         $this->get("/posts/{$post->slug}")
             ->assertOk()
