@@ -4,12 +4,13 @@ namespace App\Models;
 
 use App\Concern\Likeable;
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
 class Comment extends Model
 {
-    use Likeable;
+    use HasFactory, Likeable;
 
     /**
      * The attributes that are mass assignable.
