@@ -1,14 +1,30 @@
 <?php
 
+namespace Database\Factories;
+
 use App\Models\Comment;
 use App\Models\Post;
 use App\Models\User;
-use Faker\Generator;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(Comment::class, function (Generator $faker) {
-    return [
-        'content' => $faker->paragraph,
-        'author_id' => fn () => factory(User::class)->create()->id,
-        'post_id' => fn () => factory(Post::class)->create()->id
-    ];
-});
+class CommentFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Comment::class;
+
+    /**
+     * Define the model's default state.
+     */
+    public function definition(): array
+    {
+        return [
+            'content' => $this->faker->paragraph,
+            'author_id' => User::factory(),
+            'post_id' => Post::factory()
+        ];
+    }
+}
