@@ -1,10 +1,26 @@
 <?php
 
-use App\Models\NewsletterSubscription;
-use Faker\Generator;
+namespace Database\Factories;
 
-$factory->define(NewsletterSubscription::class, function (Generator $faker) {
-    return [
-        'email' => $faker->unique()->safeEmail
-    ];
-});
+use App\Models\NewsletterSubscription;
+use Illuminate\Database\Eloquent\Factories\Factory;
+
+class NewsletterSubscriptionFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = NewsletterSubscription::class;
+
+    /**
+     * Define the model's default state.
+     */
+    public function definition(): array
+    {
+        return [
+            'email' => $this->faker->unique()->safeEmail
+        ];
+    }
+}
