@@ -1,16 +1,30 @@
 <?php
 
+namespace Database\Factories;
+
 use App\Models\Post;
 use App\Models\User;
-use Faker\Generator;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(Post::class, function (Generator $faker) {
-    return [
-        'title' => $faker->sentence,
-        'content' => $faker->paragraph,
-        'posted_at' => now(),
-        'author_id' => function () {
-            return factory(User::class)->create()->id;
-        }
-    ];
-});
+class PostFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Post::class;
+
+    /**
+     * Define the model's default state.
+     */
+    public function definition(): array
+    {
+        return [
+            'title' => $this->faker->sentence,
+            'content' => $this->faker->paragraph,
+            'posted_at' => now(),
+            'author_id' => User::factory()
+        ];
+    }
+}
