@@ -19,7 +19,7 @@ class CreateMediaTableWithMediaLibrary extends Migration
             $table->string('file_name');
             $table->string('mime_type')->nullable();
             $table->string('disk');
-            $table->unsignedInteger('size');
+            $table->unsignedBigInteger('size');
             $table->json('manipulations');
             $table->json('custom_properties');
             $table->json('responsive_images');
